genie3.0/genefinal/analytics/requirement.txt
pandas>=1.5.0
numpy>=1.24.0
mysql-connector-python>=8.0.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.17.0
scikit-learn>=1.3.0
scipy>=1.10.0
statsmodels>=0.14.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
warnings
datetime
